merge from Develop

+++ apps/api-backend/src/controllers/webhooks/clerk/clerk-webhook.service.ts
@@ -4,9 +4,8 @@ import {
   AppUserUpdateError,
 } from '@/types/errors';
 import { UserJSON, WebhookEvent } from '@clerk/backend';
-import { IAppUserRepository } from '@domains/app-user/app-user.repository.interface';
+import { IAppUsersRepository } from '@domains/app-users/app-users.repository.interface';
 import { Inject, Injectable, Logger } from '@nestjs/common';
-import { PrismaClientManager } from '@tech-post-cast/database';
 
 @Injectable()
 export class ClerkWebhookService {
@@ -14,8 +13,7 @@ export class ClerkWebhookService {
 
   constructor(
     @Inject('AppUserRepository')
-    private readonly appUserRepository: IAppUserRepository,
-    private readonly prismaClientManager: PrismaClientManager,
+    private readonly appUserRepository: IAppUsersRepository,
   ) {}
 
   /**
@@ -67,9 +65,19 @@ export class ClerkWebhookService {
         email: userJson.email_addresses[0].email_address,
         firstName: userJson.first_name,
         lastName: userJson.last_name,
+        displayName: `${userJson.first_name} ${userJson.last_name}`,
         imageUrl: userJson.image_url,
         isActive: true,
         lastSignInAt: new Date(userJson.last_sign_in_at),
+        stripeCustomerId: undefined,
+        defaultPaymentMethodId: undefined,
+        slackWebhookUrl: undefined,
+        notificationEnabled: false,
+        rssToken: null,
+        rssEnabled: false,
+        rssCreatedAt: null,
+        rssUpdatedAt: null,
+        personalizedProgramDialogueEnabled: false,
         createdAt: new Date(userJson.created_at),
         updatedAt: new Date(userJson.updated_at),
       });
@@ -96,18 +104,34 @@ export class ClerkWebhookService {
     });
     // ユーザー更新時の処理
     try {
-      const appUser = await this.appUserRepository.update({
+      const appUser = await this.appUserRepository.findOne(userJson.id);
+      if (!appUser) {
+        this.logger.warn(`ユーザー [${userJson.id}] が見つかりません`);
+        return;
+      }
+      const updatedUser = await this.appUserRepository.update({
         id: userJson.id,
         email: userJson.email_addresses[0].email_address,
         firstName: userJson.first_name,
         lastName: userJson.last_name,
+        displayName: `${userJson.first_name} ${userJson.last_name}`,
         imageUrl: userJson.image_url,
         isActive: true,
         lastSignInAt: new Date(userJson.last_sign_in_at),
+        stripeCustomerId: appUser.stripeCustomerId,
+        defaultPaymentMethodId: appUser.defaultPaymentMethodId,
+        slackWebhookUrl: appUser.slackWebhookUrl,
+        notificationEnabled: appUser.notificationEnabled,
+        rssToken: appUser.rssToken,
+        rssEnabled: appUser.rssEnabled,
+        rssCreatedAt: appUser.rssCreatedAt,
+        rssUpdatedAt: appUser.rssUpdatedAt,
+        personalizedProgramDialogueEnabled:
+          appUser.personalizedProgramDialogueEnabled,
         createdAt: new Date(userJson.created_at),
         updatedAt: new Date(userJson.updated_at),
       });
-      this.logger.log(`ユーザー [${appUser.id}] を更新しました`);
+      this.logger.log(`ユーザー [${updatedUser.id}] を更新しました`);
     } catch (error) {
       const errorMessage = `ユーザーの更新に失敗しました`;
       this.logger.error(errorMessage, {
@@ -134,6 +158,13 @@ export class ClerkWebhookService {
     }
     // ユーザー削除時の処理
     try {
+      const user = await this.appUserRepository.findOne(userId);
+      if (!user) {
+        this.logger.warn(
+          `ユーザー [${userId}] は登録されていません。削除処理はスキップします。`,
+        );
+        return;
+      }
       await this.appUserRepository.delete(userId);
       this.logger.log(`ユーザー [${userId}] を削除しました`);
     } catch (error) {
